modulePartition
