+++ index.html
@@ -226,7 +226,7 @@
       </div>
     </section>
   </template>
-<script src="main.js"></script>
+  <script type="module" src="main.js"></script>
 <script src="function.js"></script>
 </body>
 </html>
